build: move config into package
