+++ src/professional_invoice_manager/__init__.py
@@ -1,0 +1,1 @@
+"""Professional Invoice Manager package"""
